grp-33 -modify group options - done grp-34 -split userProfile into smaller classes - done grp-47 -back and logout bug - fixed
